cleanup active admin extensions
