+++ lib/extensions/active_admin/active_admin_custom_header.rb
@@ -1,0 +1,6 @@
+module ActiveAdminCustomHeader
+  def build(*args)
+    super(*args)
+    render 'admin/custom_header'
+  end
+end
