+++ lib/extensions/active_admin/active_admin_publishable.rb
@@ -1,0 +1,11 @@
+module ActiveAdminPublishable
+  module Scopes
+    def publishable_scopes
+      scope('All', &:all)
+      scope('Draft', &:draft)
+      scope('Pending', &:pending)
+      scope('Published', &:published)
+      scope('Archived', &:archived)
+    end
+  end
+end
